move `matchVariant` tests to own file
